email: Move `OwnerInviteEmail` struct to its corresponding module
